backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
requests==2.31.0